--- a/util/file_reader_writer.cc
+++ b/util/file_reader_writer.cc
@@ -685,7 +685,7 @@ Status FilePrefetchBuffer::Prefetch(RandomAccessFileReader* reader,
 
   Slice result;
   s = reader->Read(rounddown_offset + chunk_len,
-                   static_cast<size_t>(roundup_len), &result,
+                   static_cast<size_t>(roundup_len - chunk_len), &result,
                    buffer_.BufferStart() + chunk_len);
   if (s.ok()) {
     buffer_offset_ = rounddown_offset;